% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{owned}
\alias{owned}
\title{Owned data sets}
\format{A data.frame with 18 variables:
  \code{SetID}
  \code{Number}
  \code{Variant}
  \code{Theme}
  \code{Subtheme}
  \code{Year}
  \code{Name}
  \code{Minifigs}
  \code{Pieces}
  \code{UKPrice}
  \code{USPrice}
  \code{CAPPrice}
  \code{EUPrice}
  \code{ImageURL}
  \code{Owned}
  \code{Wanted}
  \code{QtyOwned}
  \code{Rating}}
\usage{
owned
}
\description{
ids for owned data sets
}
\keyword{datasets}

